change(nr.databind): IDataType.check_value() was replaced with IDataType.isinstance_check()
This new method accepts two more parameters, "strict" and "coerce". This allows that we
employ a non-strict behavior in most scenarios (as previously) but use a strict instance
check for the UnionType.

Some IDataType implementations already used a hidden parameter to prevent coercion if not
needed, now with the "coerce" parameter this can be avoided in a supported way.